switch to new configuration with additional plugins
This patch adds the additional plugins. We now use uv from astral to
setup and install a python virtualenv when python is not installed.

We require nvm to in order to install pyright and vim-language-server.

Note: we only test against nvim 0.10.3 on homebrew. Consider using
homebrew linux to install neovim on linux hosts.